backend/requirements.txt
# ============================================================
#  Triplet Backend — Python Dependencies
#  Install: pip install -r requirements.txt
# ============================================================

# ── Web framework ────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9        # file uploads

# ── Database ─────────────────────────────────────────────────
sqlalchemy==2.0.30
asyncpg==0.29.0                # async PostgreSQL driver
psycopg2-binary==2.9.9         # sync driver (Alembic)
alembic==1.13.1                # DB migrations
pgvector==0.2.5                # pgvector SQLAlchemy type

# ── Redis + Celery ───────────────────────────────────────────
redis==5.0.4
celery==5.4.0
flower==2.0.1                  # Celery task monitor
slowapi==0.1.9                 # Redis-backed rate limiting (FastAPI)
limits==3.13.0                 # slowapi storage backend

# ── Auth & Security ──────────────────────────────────────────
python-jose[cryptography]==3.3.0   # JWT
passlib[bcrypt]==1.7.4             # password hashing
bcrypt==4.0.1                      # passlib 1.7.x compatibility; newer bcrypt breaks bcrypt backend detection
cryptography==42.0.7               # RSA signing

# ── MinIO (object storage) ───────────────────────────────────
minio==7.2.7

# ── AI / ML ──────────────────────────────────────────────────
sentence-transformers==2.7.0   # SBERT (all-MiniLM-L6-v2)
spacy==3.8.3                   # FIX: upgraded from 3.7.4 — needs typer>=0.3 <0.10 which conflicts with fastapi-cli; 3.8.x lifted that cap
torch==2.3.0                   # sentence-transformers dependency
numpy==1.26.4
typer>=0.9.0                   # FIX: explicit pin satisfies both spacy 3.8 and fastapi-cli

# ── Document parsing ─────────────────────────────────────────
pdfplumber==0.11.0             # PDF text extraction
pypdf==4.2.0                   # fallback PDF text extraction
python-docx==1.1.0             # DOCX parsing
Pillow==10.3.0                 # image handling for OCR
pytesseract==0.3.10            # OCR for image-based PDFs
PyMuPDF==1.24.9                # OCR page rendering for scanned PDFs

# ── Email ────────────────────────────────────────────────────
aiosmtplib==3.0.1              # async SMTP

# ── Validation & config ──────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1

# ── Utilities ────────────────────────────────────────────────
python-dotenv==1.0.1
httpx==0.27.0                  # async HTTP client
tenacity==8.3.0                # retry logic

# ── Testing ──────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
